Anpassungen für Snowflake

+++ USS_abbildung/DDL/uss_willibald/table_bestellung.sql
@@ -3,8 +3,8 @@
 
 CREATE TABLE uss_willibald.bestellung (
 	md_inserted_at timestamp default current_timestamp,
-	_key_bestellung int8 NOT NULL,
-	bestellungid int8 NOT NULL,
+	_key_bestellung number(20) NOT NULL,
+	bestellungid number(20) NOT NULL,
 	bestelldatum date NOT NULL,
 	wunschdatum date NOT NULL,
 	rabatt numeric(5, 2) NOT NULL
